watch-serve.sh: Add container support
for people that have a container runtime available and prefer not to
install the dependencies locally. If the dependencies are installed
locally, it will call these directly instead of using the container.
Otherwise it will check if podman is installed and use that, or docker
otherwise, build the container from the `Containerfile` and run it with
the project directory mounted.

If neither the dependencies nor a suitable container runtime can be
found, a messages is printed to stdout that tells the user what is
needed to use the script.