actions: Rewrite docs
In preparation of https://github.com/zellij-org/zellij/pull/1570 which
may land soon. This PR will document all the actions directly inside the
code, making the docs available on `docs.rs`. So instead of duplicating
the documentation here, instead give the reader some ideas and examples
of cool things to do with the actions.