Make it clear what the config directory is
The wording here was not 100% clear (to me at least), where the actual
layout directory is.
By linking to the configuration documentation, which clearly states that
the configuration directory is `.config/zellij`, and the sentence that
the layout directory is a subdirectory to that, this clearifies the
location for me.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>