Fix shell script to select a session to attach to
The format of the `zellij list-sessions` changed. To keep the colourful output as well as the additional information for the user, `--no-formatting` flag is not used. Instead, the line selected by the fuzzy finder gets trimmed to the first column (delimited by whitespace) with the session name. The no longer maintained fuzzy finder `skim` is replaced with `fzf`.